+++ droidboy/src/com/appboy/sample/FeedCategoriesFragment.java
@@ -9,6 +9,7 @@ import android.support.v4.app.DialogFragment;
 import android.util.Log;
 import android.widget.ListView;
 
+import com.appboy.Constants;
 import com.appboy.enums.CardCategory;
 
 import java.util.Arrays;
@@ -17,6 +18,7 @@ import java.util.List;
 import java.util.Locale;
 
 public class FeedCategoriesFragment extends DialogFragment {
+  private static final String TAG = String.format("%s.%s", Constants.APPBOY_LOG_TAG_PREFIX, FeedCategoriesFragment.class.getName());
   public static final String CATEGORIES_STRING= "categories";
 
   /* The activity that creates an instance of this dialog fragment must
@@ -128,8 +130,7 @@ public class FeedCategoriesFragment extends DialogFragment {
         if (i >= 0) {
           array[i] = true;
         } else {
-          Log.i("[FeedCategoriesFragment]: getBooleansFromEnumSet", "Cannot find Category for" +
-              category.toString() + "in the categories list.");
+          Log.i(TAG, "Cannot find Category for" + category.toString() + "in the categories list.");
         }
       }
       return array;